+++ tests/test_imports.py
@@ -1,3 +1,8 @@
+"""
+Test imports separately for each module and top level functions/classes. These
+tests will give hints in case a single export breaks, whereas im `import pyfar`
+would fail completely and make possible issues harder to find.
+"""
 import importlib
 
 
@@ -6,42 +11,37 @@ def test_import_importlib():
 
 
 def test_import_pyfar():
-    import pyfar
-    __all__ = [pyfar]
-    return __all__
+    import pyfar                             # noqa: F401
 
 
 def test_import_classes():
-    from pyfar import Signal
-    from pyfar import TimeData
-    from pyfar import FrequencyData
+    from pyfar import Signal                 # noqa: F401
+    from pyfar import TimeData               # noqa: F401
+    from pyfar import FrequencyData          # noqa: F401
 
-    from pyfar import Coordinates
-    from pyfar import Orientations
+    from pyfar import Coordinates            # noqa: F401
+    from pyfar import Orientations           # noqa: F401
 
-    from pyfar import FilterSOS
-    from pyfar import FilterFIR
-    from pyfar import FilterIIR
-
-    __all__ = [
-        Signal,
-        TimeData,
-        FrequencyData,
-        Coordinates,
-        Orientations,
-        FilterSOS,
-        FilterFIR,
-        FilterIIR
-    ]
-    return __all__
+    from pyfar import FilterSOS              # noqa: F401
+    from pyfar import FilterFIR              # noqa: F401
+    from pyfar import FilterIIR              # noqa: F401
 
 
 def test_import_submodules():
-    import pyfar
-    assert pyfar.dsp
-    assert pyfar.dsp.fft
-    assert pyfar.dsp.filter
-    assert pyfar.io
-    assert pyfar.samplings
-    assert pyfar.plot
-    assert pyfar.signals
+    from pyfar import dsp                    # noqa: F401
+    from pyfar.dsp import fft                # noqa: F401
+    from pyfar.dsp import filter             # noqa: F401
+    from pyfar import io                     # noqa: F401
+    from pyfar import plot                   # noqa: F401
+    from pyfar import samplings              # noqa: F401
+    from pyfar import signals                # noqa: F401
+    from pyfar.signals import files          # noqa: F401
+
+
+def test_import_functions():
+    from pyfar import add                    # noqa: F401
+    from pyfar import subtract               # noqa: F401
+    from pyfar import multiply               # noqa: F401
+    from pyfar import divide                 # noqa: F401
+    from pyfar import power                  # noqa: F401
+    from pyfar import matrix_multiplication  # noqa: F401